pffimport: fix i586 build error
tools/pffimport.cpp:280:44: error: cannot convert 'size_t*' {aka 'unsigned
int*'} to 'size64_t*' {aka 'long long unsigned int*'}
 if (libpff_attachment_get_data_size(atx, &asize, nullptr) < 1)

+++ tools/pffimport.cpp
@@ -271,7 +271,7 @@ static void az_recordset(unsigned int depth, libpff_record_set_t *rset)
 static int do_attach(unsigned int depth, ATTACHMENT_CONTENT *atc, libpff_item_t *atx)
 {
 	int atype = 0;
-	size_t asize = 0;
+	uint64_t asize = 0;
 
 	if (libpff_attachment_get_type(atx, &atype, nullptr) < 1)
 		throw "PF-1012";
